cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multiboot"
version = "0.9.0"
description = "Upload GBA multiboot payloads via Raspberry Pi Pico"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "James Smith", email = "james@loopj.com" }]
dependencies = ["pyserial", "rich"]

[project.scripts]
multiboot = "multiboot.cli:main"

[tool.ruff.lint]
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "D",  # pydocstyle
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "Q",  # flake8-quotes
]
ignore = [
    "E501", # line too long, handled by formatter
    "C901", # mccabe complexity, disabled
]

[tool.ruff.lint.isort]
known-first-party = ["multiboot"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
typeCheckingMode = "strict"
